packages/ml/flaskapi/requirements.txt
Flask==2.0.2
colorama==0.4.4
click==8.0.3
importlib-metadata==4.8.2
itsdangerous==2.1.0
Jinja2==3.0.3
Markupsafe==2.0.1
typing-extensions==4.0.1
Werkzeug==2.0.3
zipp==3.6.0
gunicorn==20.1.0
PyPDF2==3.0.1
scikit-learn==1.2.2
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
ipython==8.15.0